clipper-video/backend/requirements.txt
Django==4.2.28
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
yt-dlp
